requirement.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
lxml
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
beautifulsoup4
streamlit
langchain-ollama
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
arxiv
wikipedia
duckduckgo-search
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
PyPDF2
gradio
crewai
crewai_tools
pinecone-client
pinecone-text
pinecone-notebooks
neo4j==5.14
langgraph
langsmith
cassio
tiktoken
cassio